Do not parse propers for '__symbols__' node
It's a special Device Tree Compiler node which should not be treated
like a proper node. I ran into this because I had a DTB with a
property 'clocks' which was a symbol/alias for a string. This caused
dtb.zig to crash since it wasn't the expected value for clocks.

Signed-off-by: Ivan-Velickovic <i.velickovic@unsw.edu.au>